+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `Object::singleton_class`.
 - `Proc::new`, `Proc::arity`, and `Proc::is_lambda`.
 - `superclass` and `name` methods for `RClass`.
+- `scan_args::check_arity`
 
 ### Changed
 - `RArray::cat`, `RArray::from_slice`, and `gc::mark_slice` will accept a